requirements.txt
python-telegram-bot[job-queue,webhooks]==21.7
tzdata==2024.1
